build: fix signed/unsigned comparisons flagged by MSVC C4018 / -Wsign-compare
Cast the sizeof-based OIDC_CFG_OPTIONS_SIZE, OIDC_METRICS_BUCKET_NUM and
OIDC_CONTENT_HANDLER_MAX macros to int, use size_t/unsigned loop indexes
where the bound is a size_t, and cast at the remaining mixed-signedness
comparison sites (curl proxy auth_type, aud size vs nelts, claim-size
warning, file-size check, shm value length).

+++ src/cfg/cfg.h
@@ -149,7 +149,7 @@ typedef struct oidc_pass_user_info_as_t {
 /* actions to be taken on access token / userinfo refresh error */
 typedef enum { OIDC_ON_ERROR_502 = 0, OIDC_ON_ERROR_LOGOUT = 1, OIDC_ON_ERROR_AUTH = 2 } oidc_on_error_action_t;
 
-#define OIDC_CFG_OPTIONS_SIZE(options) sizeof(options) / sizeof(oidc_cfg_option_t)
+#define OIDC_CFG_OPTIONS_SIZE(options) ((int)(sizeof(options) / sizeof(oidc_cfg_option_t)))
 
 typedef struct oidc_provider_t oidc_provider_t;
 typedef struct oidc_cfg_t oidc_cfg_t;
